Added frontend styles
I got some practice building a state machine for a page using bootstrap 5.1 floating labels. This code is mainly used to change the styles dependent on the width of the window and whether the TextAreas on the page have scrollbars as well as the position of the scrollbar if so. 

There are two real states (SmallWindow and RegularWindow), a Context class to hold the states and perform the styling actions, and a Factory which supplies the correct state to the Context and also caches the objects to save memory. I just learned most of this as I'm primarily a C# developer and JS is a little different.

I need to add more comments for how to build a page using this script.